requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
dj-database-url==2.1.0
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.10.0
django-redis==5.3.0
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
pytz==2023.3
redis==5.0.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
whitenoise==6.5.0
zope.interface==6.0
